Remove AllGatherTestWithTimeout
This test previously did a thread sleep before launching the allgather operation, and then waited on the work object. Since the sleep was done before the work object was created, it did not affect the allgather call, and thus, did not test work-level timeouts as intended.

I am removing this test for now. In the future we can add this test back, but would need to somehow inject a `cudaSleep` call before the  allgather (so the collective operation itself is delayed). This may require overriding the `ProcessGroupNCCL::collective`, so it's a bit more heavy-weight.

In the meantime, we can remove this test - work-level timeouts are still thoroughly tested with Gloo.

Differential Revision: [D22702291](https://our.internmc.facebook.com/intern/diff/D22702291/)

ghstack-source-id: 108370178
Pull Request resolved: https://github.com/pytorch/pytorch/pull/41945