Simple caching allocator for CPU.
Summary:
This PR introduces a simple CPU caching allocator. This is specifically
intended for mobile use cases and for inference. There is nothing
specific to the implementation that can prevent it from other use cases,
however its simplicity may not be suitable everywhere.
It simply tracks allocation by sizes and relies on deterministic
repeatable behavior where allocation of same sizes are made on every
inference.
Thus after the first allocation when the pointer is returned, instead of
returning it to system, allocator caches it for subsequent use.
Memory is freed automatically at the end of the process, or it can be
explicitly freed.
This is enabled at the moment in DefaultMobileCPUAllocator only.

Test Plan:
android test: cpu_caching_allocator_test

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 3a5ca70002711d59896e2e71c42185db5d4a4c5c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42006